configs/cron: added FIXME to zfs-auto-snapshot
This two-at-once problem has actually happened, although only once.  The
result in an odd email from cron saying that zfs-auto-snaphshot couldn't
find a snapshot to destroy.  When it happened, I was in the middle of doing
both the nighly amanda backups and a test run of zfs-replication... so
that's under abnormally high load.