net/tc-rules: overhaul settings for all 3 cah hosts, v1
This commit adds specific rules files for aragorn and srv-router2, to
go along with the main rules for srv-router1.  We're attempting to
limit the likelyhood of interface overruns on srv-router1 because of
shared WAN interfaces with aragorn and srv-router2 (which really only
have their own WAN interfaces for emergency access) by capping usage
on aragorn and srv-router2 and lowering MEDIUM guaranteed rate on
srv-router1.

Had to tweak the apply-tc-rules nm-dispatcher to prefer
SITE-HOSTNAME-INTNAME-tc-rules if present, too.