net/check-gateway: skip single-ping ip_is_reachable() test
Using a single ping to verify reachability of the gateway in
ip_is_reachable() was never a good idea, because a single ping *can*
get lost here and there even on a good gateway.  Now that we're
relying on check-gateway (and have tons of logfiles to look at), it's
clearly causing some false-positives and unnecesary interface
swapping.

This commit removes the call to ip_is_reachable(), and dives straight
into ip_is_good_gw(), which uses pingenator to spam a bunch of hosts
(and returns successfully if it gets any replies from any of them).
If the gw is actually unreachable, ip_is_good_gw() works just fine and
doesn't take any longer.

The whole ip_is_reachable() thing is really a historical artifact,
first from before implementation of pingenator (the workhorse of
ip_is_good_gw()), then because we were initially only calling
pingenator if not --dry-run.  Once I'd established enough trust in
pingenator not leaving the system in a wonky state (it *does* add
potential routes to the system to test them), and enabled calling it
even when --dry-run, this just became a waste of code.

Juuuust in case I change my mind, I've left ip_is_reachable() in place
and put a big FIXME in there about the false-positives.  But we no
longer call it.