net/check-gateway: check all possible gateways every time
This commit reorganizes the code quite a bit so that we can send
alerts if a backup WAN interface isn't working.  We now check all
configured interfaces in a loop, setting variables to keep track of
which ones work and which ones don't, then set a new gateway if
needed.  At the very end, alerts are emailed similar to the the
existing "gateway modified" email.