net: add $lock check to check-gateway and check-wireguard
Wow, I'm surprised this didn't cause problems.  We were creating a
lockfile in these two scripts, and checking for it appropriately in
mail-and-purge-logfile, but we never added code to the scripts
themselves to check it.  It wasn't hurting anything in
check-wireguard, because we run that via cron once per minute (and it
just takes a second or two to run).  I'm suprised I didn't notice
problems with check-gateway, though... that one has potentially slow
runtimes (e.g., more than 1 minute if a dhcp interfaces tries to raise
and fails), and we run it every 10 seconds.